requirements.txt
pyqt==5.9.2